requirements.txt
cycler==0.10.0
decorator==4.4.0
kiwisolver==1.1.0
matplotlib==3.0.3
networkx==2.3
numpy==1.16.3
pyparsing==2.4.0
python-dateutil==2.8.0
python-Levenshtein==0.12.0
six==1.12.0
